[ghstack-poisoned]

+++ caffe2/core/blob_serialization.h
@@ -38,7 +38,7 @@ CAFFE2_API void SerializeBlob(
 /**
  * @brief Convenience function to serialize a blob to a string.
  *
- * This is a conveinence function to serialize small Blobs that produce
+ * This is a convenience function to serialize small Blobs that produce
  * manageable serialized strings. To serialize big blobs such as
  * large sparse tensors, use the fully-functional interface in
  * blob_serializer_base.h.
@@ -133,7 +133,7 @@ class CAFFE2_API TensorDeserializer : public BlobDeserializerBase {
    * we have to call Deserialize multiple times to get the complete deserialized
    * Tensor, each call will fill part of the Tensor given the segment begin and
    * end information in proto, therefore we have to pass in the Tensor pointer
-   * rather than create a new Tensor everytime.
+   * rather than create a new Tensor every time.
    *
    * Precondition: Tensor must be initialized
    */
@@ -162,7 +162,7 @@ inline void CopyToProtoAsIs(
       "The source type and dest type cannot be copied as-is. Did "
       "you mean CopyToProtoWithCast?");
   field->Reserve(size);
-  for (int i = 0; i < size; ++i) {
+  for (size_t i = 0; i < size; ++i) {
     field->Add(0);
   }
   context->template CopyToCPU<SrcType>(
@@ -183,7 +183,7 @@ inline void CopyToProtoWithCast(
   context->template CopyToCPU<SrcType>(size, src, buffer.get());
   context->FinishDeviceComputation();
   field->Reserve(size);
-  for (int i = 0; i < size; ++i) {
+  for (size_t i = 0; i < size; ++i) {
     field->Add(static_cast<DstType>(buffer[i]));
   }
 }
@@ -214,7 +214,7 @@ inline void CopyFromProtoWithCast(
   // CPUContext. Remove it if it is performance critical.
   unique_ptr<DstType[]> buffer(new DstType[size]);
   const SrcType* src = field.data();
-  for (int i = 0; i < size; ++i) {
+  for (size_t i = 0; i < size; ++i) {
     buffer[i] = static_cast<DstType>(src[i]);
   }
   context->template CopyFromCPU<DstType>(size, buffer.get(), dst);
